+++ projects/demo/src/app/pages/examples/examples.module.ts
@@ -32,8 +32,9 @@ import {
   ButtonViewComponent,
 } from './custom-edit-view/basic-example-button-view.component';
 import {BasicExampleCustomActionsItemComponent} from './custom-edit-view/basic-example-custom-actions-item.component';
-import {BasicExampleRowSelectionComponent} from './various/row-selection/basic-example-row-selection.component';
+import {BasicExampleRowSelectionComponent} from './various/basic-example-row-selection.component';
 import {ColumnHideExampleComponent} from './filter/column-hide-example.component';
+import {BasicExampleRowHeadingsComponent} from "./various/basic-example-row-headings.component";
 
 const EXAMPLES_COMPONENTS = [
   AdvancedExampleFiltersComponent,
@@ -47,6 +48,7 @@ const EXAMPLES_COMPONENTS = [
   BasicExampleMultiSelectComponent,
   BasicExampleSourceComponent,
   BasicExampleRowSelectionComponent,
+  BasicExampleRowHeadingsComponent,
   CustomEditorComponent,
   CustomRenderComponent,
   CustomFilterComponent,